refactor: code rehabilitation with new devices/features
