+++ server/services/netatmo/lib/index.js
@@ -10,6 +10,9 @@ const { getConfiguration } = require('./netatmo.getConfiguration');
 const { getStatus } = require('./netatmo.getStatus');
 const { saveStatus } = require('./netatmo.saveStatus');
 const { saveConfiguration } = require('./netatmo.saveConfiguration');
+const { convertDeviceEnergy } = require('./device/netatmo.convertDeviceEnergy');
+const { convertDeviceWeather } = require('./device/netatmo.convertDeviceWeather');
+const { convertDeviceNotSupported } = require('./device/netatmo.convertDeviceNotSupported');
 const { discoverDevices } = require('./netatmo.discoverDevices');
 const { loadDevices } = require('./netatmo.loadDevices');
 const { loadDeviceDetails } = require('./netatmo.loadDeviceDetails');
@@ -19,14 +22,14 @@ const { pollRefreshingToken } = require('./netatmo.pollRefreshingToken');
 const { pollRefreshingValues } = require('./netatmo.pollRefreshingValues');
 const { setValue } = require('./netatmo.setValue');
 const { updateValues } = require('./netatmo.updateValues');
-const { updateNAPlug } = require('./device/netatmo.updateNAPlug');
-const { updateNATherm1 } = require('./device/netatmo.updateNATherm1');
-const { updateNRV } = require('./device/netatmo.updateNRV');
-const { updateNAMain } = require('./device/netatmo.updateNAMain');
-const { updateNAModule1 } = require('./device/netatmo.updateNAModule1');
-const { updateNAModule2 } = require('./device/netatmo.updateNAModule2');
-const { updateNAModule3 } = require('./device/netatmo.updateNAModule3');
-const { updateNAModule4 } = require('./device/netatmo.updateNAModule4');
+const { updateNAPlug } = require('./update/netatmo.updateNAPlug');
+const { updateNATherm1 } = require('./update/netatmo.updateNATherm1');
+const { updateNRV } = require('./update/netatmo.updateNRV');
+const { updateNAMain } = require('./update/netatmo.updateNAMain');
+const { updateNAModule1 } = require('./update/netatmo.updateNAModule1');
+const { updateNAModule2 } = require('./update/netatmo.updateNAModule2');
+const { updateNAModule3 } = require('./update/netatmo.updateNAModule3');
+const { updateNAModule4 } = require('./update/netatmo.updateNAModule4');
 
 const { STATUS, SCOPES } = require('./utils/netatmo.constants');
 const buildScopesConfig = require('./utils/netatmo.buildScopesConfig');
@@ -63,6 +66,9 @@ NetatmoHandler.prototype.getRefreshToken = getRefreshToken;
 NetatmoHandler.prototype.refreshingTokens = refreshingTokens;
 NetatmoHandler.prototype.getConfiguration = getConfiguration;
 NetatmoHandler.prototype.saveConfiguration = saveConfiguration;
+NetatmoHandler.prototype.convertDeviceEnergy = convertDeviceEnergy;
+NetatmoHandler.prototype.convertDeviceWeather = convertDeviceWeather;
+NetatmoHandler.prototype.convertDeviceNotSupported = convertDeviceNotSupported;
 NetatmoHandler.prototype.discoverDevices = discoverDevices;
 NetatmoHandler.prototype.loadDevices = loadDevices;
 NetatmoHandler.prototype.loadDeviceDetails = loadDeviceDetails;
