+++ server/services/netatmo/lib/netatmo.discoverDevices.js
@@ -1,7 +1,6 @@
 const logger = require('../../../utils/logger');
 const { ServiceNotConfiguredError } = require('../../../utils/coreErrors');
-const { STATUS } = require('./utils/netatmo.constants');
-const { convertDevice } = require('./device/netatmo.convertDevice');
+const { STATUS, SUPPORTED_CATEGORY_TYPE } = require('./utils/netatmo.constants');
 
 /**
  * @description Discover Netatmo cloud devices.
@@ -26,20 +25,36 @@ async function discoverDevices() {
     logger.error('Unable to load Netatmo devices', e);
   }
   if (devicesNetatmo.length > 0) {
-    this.discoveredDevices = devicesNetatmo.map((device) => ({
-      ...convertDevice(device),
-      service_id: this.serviceId,
-      deviceNetatmo: device,
-    }));
+    this.discoveredDevices = devicesNetatmo.map((device) => {
+      let discoveredDevice;
+      switch (device.categoryAPI) {
+        case SUPPORTED_CATEGORY_TYPE.ENERGY: {
+          discoveredDevice = this.convertDeviceEnergy(device);
+          break;
+        }
+        case SUPPORTED_CATEGORY_TYPE.WEATHER: {
+          discoveredDevice = this.convertDeviceWeather(device);
+          break;
+        }
+        default:
+          discoveredDevice = this.convertDeviceNotSupported(device);
+          break;
+      }
+      return {
+        ...discoveredDevice,
+        service_id: this.serviceId,
+        deviceNetatmo: device,
+      };
+    });
     const discoveredDevices = this.discoveredDevices.filter((device) => {
       const existInGladys = this.gladys.stateManager.get('deviceByExternalId', device.external_id);
       return existInGladys === null;
     });
-    this.saveStatus({ statusType: STATUS.CONNECTED, message: null });
+    await this.saveStatus({ statusType: STATUS.CONNECTED, message: null });
     logger.debug(`${discoveredDevices.length} new Netatmo devices found`);
     return discoveredDevices;
   }
-  this.saveStatus({ statusType: STATUS.CONNECTED, message: null });
+  await this.saveStatus({ statusType: STATUS.CONNECTED, message: null });
   logger.debug('No devices found');
   return [];
 }
