+++ server/services/netatmo/lib/update/netatmo.updateNAModule3.js
@@ -1,47 +1,47 @@
 const { EVENTS } = require('../../../../utils/constants');
 const logger = require('../../../../utils/logger');
-const { readValues } = require('./netatmo.deviceMapping');
+const { readValues } = require('../device/netatmo.deviceMapping');
 
 /**
- * @description Save values of valves NRV.
+ * @description Save values of Smart Rain Gauge Module NAModule3.
  * @param {object} deviceGladys - Device object in Gladys.
  * @param {object} deviceNetatmo - Device object coming from the Netatmo API.
  * @param {string} externalId - Device identifier in gladys.
- * @example updateNRV(deviceGladys, deviceNetatmo, externalId);
+ * @example updateNAModule3(deviceGladys, deviceNetatmo, externalId);
  */
-async function updateNRV(deviceGladys, deviceNetatmo, externalId) {
-  const { room } = deviceNetatmo;
+async function updateNAModule3(deviceGladys, deviceNetatmo, externalId) {
+  const { dashboard_data: dashboardData } = deviceNetatmo;
   try {
     deviceGladys.features
       .filter((feature) => feature.external_id === `${externalId}:battery_percent`)
       .forEach((feature) => {
         this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
-          device_feature_external_id: feature.external_id,
-          state: readValues[feature.category][feature.type](deviceNetatmo.battery_state),
+          device_feature_external_id: `${externalId}:battery_percent`,
+          state: readValues[feature.category][feature.type](deviceNetatmo.battery_percent),
         });
       });
     deviceGladys.features
-      .filter((feature) => feature.external_id === `${externalId}:therm_measured_temperature`)
+      .filter((feature) => feature.external_id === `${externalId}:rain`)
       .forEach((feature) => {
         this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
           device_feature_external_id: feature.external_id,
-          state: readValues[feature.category][feature.type](room.therm_measured_temperature),
+          state: readValues[feature.category][feature.type](deviceNetatmo.rain),
         });
       });
     deviceGladys.features
-      .filter((feature) => feature.external_id === `${externalId}:therm_setpoint_temperature`)
+      .filter((feature) => feature.external_id === `${externalId}:sum_rain_1`)
       .forEach((feature) => {
         this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
           device_feature_external_id: feature.external_id,
-          state: readValues[feature.category][feature.type](room.therm_setpoint_temperature),
+          state: readValues[feature.category][feature.type](deviceNetatmo.sum_rain_1),
         });
       });
     deviceGladys.features
-      .filter((feature) => feature.external_id === `${externalId}:open_window`)
+      .filter((feature) => feature.external_id === `${externalId}:sum_rain_24`)
       .forEach((feature) => {
         this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
           device_feature_external_id: feature.external_id,
-          state: readValues[feature.category][feature.type](room.open_window),
+          state: readValues[feature.category][feature.type](dashboardData.sum_rain_24),
         });
       });
     deviceGladys.features
@@ -52,20 +52,11 @@ async function updateNRV(deviceGladys, deviceNetatmo, externalId) {
           state: readValues[feature.category][feature.type](deviceNetatmo.rf_strength),
         });
       });
-    deviceGladys.features
-      .filter((feature) => feature.external_id === `${externalId}:heating_power_request`)
-      .forEach((feature) => {
-        const value = room.heating_power_request > 0;
-        this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
-          device_feature_external_id: feature.external_id,
-          state: readValues[feature.category][feature.type](value),
-        });
-      });
   } catch (e) {
-    logger.error('deviceGladys NRV: ', deviceGladys.name, 'error: ', e);
+    logger.error('deviceGladys NAModule3: ', deviceGladys.name, 'error: ', e);
   }
 }
 
 module.exports = {
-  updateNRV,
+  updateNAModule3,
 };
