+++ server/services/netatmo/lib/update/netatmo.updateNRV.js
@@ -1,47 +1,47 @@
 const { EVENTS } = require('../../../../utils/constants');
 const logger = require('../../../../utils/logger');
-const { readValues } = require('./netatmo.deviceMapping');
+const { readValues } = require('../device/netatmo.deviceMapping');
 
 /**
- * @description Save values of Smart Rain Gauge Module NAModule3.
+ * @description Save values of valves NRV.
  * @param {object} deviceGladys - Device object in Gladys.
  * @param {object} deviceNetatmo - Device object coming from the Netatmo API.
  * @param {string} externalId - Device identifier in gladys.
- * @example updateNAModule3(deviceGladys, deviceNetatmo, externalId);
+ * @example updateNRV(deviceGladys, deviceNetatmo, externalId);
  */
-async function updateNAModule3(deviceGladys, deviceNetatmo, externalId) {
-  const { dashboard_data: dashboardData } = deviceNetatmo;
+async function updateNRV(deviceGladys, deviceNetatmo, externalId) {
+  const { room } = deviceNetatmo;
   try {
     deviceGladys.features
       .filter((feature) => feature.external_id === `${externalId}:battery_percent`)
       .forEach((feature) => {
         this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
-          device_feature_external_id: `${externalId}:battery_percent`,
-          state: readValues[feature.category][feature.type](deviceNetatmo.battery_percent),
+          device_feature_external_id: feature.external_id,
+          state: readValues[feature.category][feature.type](deviceNetatmo.battery_state),
         });
       });
     deviceGladys.features
-      .filter((feature) => feature.external_id === `${externalId}:rain`)
+      .filter((feature) => feature.external_id === `${externalId}:therm_measured_temperature`)
       .forEach((feature) => {
         this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
           device_feature_external_id: feature.external_id,
-          state: readValues[feature.category][feature.type](deviceNetatmo.rain),
+          state: readValues[feature.category][feature.type](room.therm_measured_temperature),
         });
       });
     deviceGladys.features
-      .filter((feature) => feature.external_id === `${externalId}:sum_rain_1`)
+      .filter((feature) => feature.external_id === `${externalId}:therm_setpoint_temperature`)
       .forEach((feature) => {
         this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
           device_feature_external_id: feature.external_id,
-          state: readValues[feature.category][feature.type](deviceNetatmo.sum_rain_1),
+          state: readValues[feature.category][feature.type](room.therm_setpoint_temperature),
         });
       });
     deviceGladys.features
-      .filter((feature) => feature.external_id === `${externalId}:sum_rain_24`)
+      .filter((feature) => feature.external_id === `${externalId}:open_window`)
       .forEach((feature) => {
         this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
           device_feature_external_id: feature.external_id,
-          state: readValues[feature.category][feature.type](dashboardData.sum_rain_24),
+          state: readValues[feature.category][feature.type](room.open_window),
         });
       });
     deviceGladys.features
@@ -52,11 +52,20 @@ async function updateNAModule3(deviceGladys, deviceNetatmo, externalId) {
           state: readValues[feature.category][feature.type](deviceNetatmo.rf_strength),
         });
       });
+    deviceGladys.features
+      .filter((feature) => feature.external_id === `${externalId}:heating_power_request`)
+      .forEach((feature) => {
+        const value = room.heating_power_request > 0;
+        this.gladys.event.emit(EVENTS.DEVICE.NEW_STATE, {
+          device_feature_external_id: feature.external_id,
+          state: readValues[feature.category][feature.type](value),
+        });
+      });
   } catch (e) {
-    logger.error('deviceGladys NAModule3: ', deviceGladys.name, 'error: ', e);
+    logger.error('deviceGladys NRV: ', deviceGladys.name, 'error: ', e);
   }
 }
 
 module.exports = {
-  updateNAModule3,
+  updateNRV,
 };
