+++ server/test/services/netatmo/lib/device/netatmo.deviceMapping.test.js
@@ -90,7 +90,7 @@ describe('Netatmo device mapping', () => {
       expect(mappingFunction(valueFromDevice)).to.equal(10);
       expect(mappingFunction(valueFromDeviceFloat)).to.equal(10);
     });
-    
+
     it('should correctly transform PRECIPITATION_SENSOR.DECIMAL value from Netatmo to Gladys', () => {
       const valueFromDevice = 1;
       const valueFromDeviceFloat = 1.5;
