+++ server/test/services/netatmo/lib/update/netatmo.updateNAMain.test.js
