+++ server/test/services/netatmo/lib/update/netatmo.updateNAModule1.test.js
