+++ server/test/services/netatmo/lib/update/netatmo.updateNAModule2.test.js
