+++ server/test/services/netatmo/lib/update/netatmo.updateNAModule3.test.js
