+++ server/test/services/netatmo/lib/update/netatmo.updateNAModule4.test.js
