+++ server/test/services/netatmo/lib/update/netatmo.updateNAPlug.test.js
