+++ server/test/services/netatmo/lib/update/netatmo.updateNATherm1.test.js
