+++ server/test/services/netatmo/netatmo.loadDevices.mock.test.json
@@ -37,7 +37,8 @@
       "type": "garage",
       "module_ids": ["01:00:00:xx:xx:xx"]
     },
-    "plug": null
+    "plug": null,
+    "categoryAPI": "Energy"
   },
   {
     "id": "04:00:00:xx:xx:xx",
@@ -133,7 +134,8 @@
       },
       "udp_conn": true,
       "last_plug_seen": 1600000000
-    }
+    },
+    "categoryAPI": "Energy"
   },
   {
     "id": "70:ee:50:yy:yy:yy",
@@ -148,7 +150,8 @@
     "last_status_store": 1601000000,
     "home": "5e1xxxxxxxxxxxxxxxxx",
     "modules_bridged": ["09:00:00:xx:xx:xx", "09:00:00:yy:yy:yy", "09:00:00:zz:zz:zz"],
-    "plug": null
+    "plug": null,
+    "categoryAPI": "Energy"
   },
   {
     "id": "09:00:00:xx:xx:xx",
@@ -187,7 +190,8 @@
       "setup_date": 1581000000,
       "room_id": "2812958323",
       "modules_bridged": ["09:00:00:xx:xx:xx", "09:00:00:yy:yy:yy", "09:00:00:zz:zz:zz"]
-    }
+    },
+    "categoryAPI": "Energy"
   },
   {
     "id": "70:ee:50:jj:jj:jj",
